[#1584] blobovniczatree: move public methods to different files
